requirements.txt
# Voz: reconocimiento y síntesis offline
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.14
gTTS>=2.3.0
pygame>=2.5.0
pydub>=0.25.1

# Utilidades del sistema
platformdirs>=4.0.0

# Para icono en la bandeja del sistema
pystray>=0.19.4
Pillow>=10.0.0  # necesario para el icono

# Para tareas periódicas
schedule>=1.2.1

# Para IA conversacional con Gemini
google-generativeai>=0.3.0